backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.13.2
psycopg[binary]==3.2.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
strawberry-graphql[fastapi]==0.246.1
python-multipart==0.0.9
